+++ src/main.rs
@@ -3,6 +3,8 @@ use runnel::RunnelIoeBuilder;
 use std::io::Write;
 
 fn main() {
+    memx_cdy::memx_init(); // fast mem operation.
+    //
     let mut env_args: Vec<String> = std::env::args().collect();
     let _program = env_args.remove(0);
     let program = env!("CARGO_PKG_NAME");