+++ CHANGELOG.md
@@ -2,6 +2,13 @@ aki-gsub TBD
 ===
 Unreleased changes. Release notes have not yet been written.
 
+0.1.31 (2021-06-24)
+=====
+
+* add `memx_cdy::memx_init(); // fast mem operation.`
+* rewite TARGET_EXE_PATH with `env!("CARGO_BIN_EXE_aki-gsub")`
+* bug fix: `#[cfg(feature = "debian_build")]`
+
 0.1.30 (2021-06-03)
 =====